+++ client.go
@@ -15,14 +15,7 @@ type Client struct {
 
 // NewClient creates a Client with http.DefaultClient.
 // If provided endpoint is not valid, an error is returned.
-func NewClient(endpoint string) (*Client, error) {
-
-	return NewCustomClient(endpoint, http.DefaultClient, make(map[string]string))
-}
-
-// NewCustomClient allows customization of http.Client and headers used to make RPC calls.
-// If provided endpoint is not valid, an error is returned.
-func NewCustomClient(endpoint string, httpClient *http.Client, headers map[string]string) (*Client, error) {
+func NewClient(endpoint string, opts ...Option) (*Client, error) {
 
 	// Parse Endpoint URL
 	endpointUrl, err := url.Parse(endpoint)
@@ -30,23 +23,25 @@ func NewCustomClient(endpoint string, httpClient *http.Client, headers map[strin
 		return nil, fmt.Errorf("invalid endpoint url: %w", err)
 	}
 
-	codec := NewCodec(endpointUrl, httpClient, headers)
+	codec := NewCodec(endpointUrl, http.DefaultClient)
 
 	c := &Client{
 		codec:  codec,
 		Client: rpc.NewClientWithCodec(codec),
 	}
 
+	// Apply options
+	for _, opt := range opts {
+		opt(c)
+	}
+
 	return c, nil
 }
 
-// UserAgent returns currently configured User-Agent header that will be sent to remote server on every RPC call.
-func (c *Client) UserAgent() string {
-	return c.codec.userAgent
-}
+// NewCustomClient allows customization of http.Client used to make RPC calls.
+// If provided endpoint is not valid, an error is returned.
+// Deprecated: prefer using NewClient with HttpClient Option
+func NewCustomClient(endpoint string, httpClient *http.Client) (*Client, error) {
 
-// SetUserAgent allows customization to User-Agent header.
-// If set to an empty string, User-Agent header will be sent with an empty value.
-func (c *Client) SetUserAgent(ua string) {
-	c.codec.userAgent = ua
+	return NewClient(endpoint, HttpClient(httpClient))
 }